Backbone of the bot has been made.
Test Module needs to be re-named at the file level.
This module will test for emoji spam which at the moment is set to 3 or more emojis in on message.
Added Checks.py file that allows us to do permissions checks on commands easily.
Added .sh file that will run the bot forever if there is an error it will restart the bot after 5 seconds.
requirements.txt file has been made
All of the bot is in the bot folder so that we can have seperate functions like the website in the same github

IMPORTANT: Need to make an example config.json file because at the moment there is none and the bot will fail without that housing the key and description of the bot.